libModulate
Implemented the basics of the library. Using the sample code from RedFox0x20/Modulators a more OOP oriented solution was produced. This library serves to be a general purpose Modulator and Demodulator providing a full suite necessary to encode and decode data for various modes using a variety of bit manipulation methods and modulation methods.