FSK Demodulation for non mod 2 bps
When bits per symbol was a non mod2 value, error were occuring within some of
the offset dependant methods. This has been resolved - Some characteristics this
leaves is situations such as where `bps = 3` and `num data bytes = 4` then there
will be an extra bit produced by the demodulation as `32 % 3 = 1`, the extra bit
will be of zero value, this could be removed using a mod2 check however it is
uncertain as to whether this would be a positive or negative change as it could
lead to the loss of bits in other cases if implemented incorrectly.

Updated FourierTransform DFT method to return the average power of a frequency
over the given window length instead of providing the sum of the power. This
should more accurately represent test frequencies within a given range. There
may be an issue caused by this where a small number of outliers could lead to
the symbol being misread, particularly at lower Signal to Noise Ratios however
this will be revealed throughout testing.