Symbols are better than bits!
Changes for #5

Moved away from converting between bits within the library for the Modulate and
Demodulate methods. This meant removing the idea of Bits Per Symbol and instead
having the library user provide a fixed value for the number of frequency levels
that will be used. By combining this with the Symbol manipulation methods made
available through the `Assigner` classes, allows for more complex operations to
be performed relatively simply such as the example code where there is a
dedicated symbol for repeated data.